+++ scripts/dev-bundle-tool-package.js
@@ -15,7 +15,7 @@ var packageJson = {
     "node-gyp": "3.5.0",
     "node-pre-gyp": "0.6.33",
     "meteor-babel": "0.19.1",
-    "meteor-promise": "0.8.0",
+    "meteor-promise": "0.8.2",
     fibers: "1.0.15",
     promise: "7.1.1",
     // So that Babel 6 can emit require("babel-runtime/helpers/...") calls.